refactor: implement suggestion
